+++ contracts/beacon/Voucher.sol
@@ -5,6 +5,7 @@ pragma solidity ^0.8.20;
 
 import {IexecLibOrders_v5} from "@iexec/poco/contracts/libs/IexecLibOrders_v5.sol";
 import {IexecPoco1} from "@iexec/poco/contracts/modules/interfaces/IexecPoco1.v8.sol";
+import {IexecPocoBoost} from "@iexec/poco/contracts/modules/interfaces/IexecPocoBoost.sol";
 import {OwnableUpgradeable} from "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IVoucherHub} from "../IVoucherHub.sol";
@@ -27,6 +28,21 @@ contract Voucher is OwnableUpgradeable, IVoucher {
         uint256 _type;
         mapping(address => bool) _authorizedAccounts;
         mapping(bytes32 dealId => uint256) _sponsoredAmounts;
+    }
+
+    modifier onlyAuthorized() {
+        VoucherStorage storage $ = _getVoucherStorage();
+        require(
+            msg.sender == owner() || $._authorizedAccounts[msg.sender],
+            "Voucher: sender is not authorized"
+        );
+        _;
+    }
+
+    modifier onlyNotExpired() {
+        VoucherStorage storage $ = _getVoucherStorage();
+        require(block.timestamp < $._expiration, "Voucher: voucher is expired");
+        _;
     }
 
     /// @custom:oz-upgrades-unsafe-allow constructor
@@ -91,31 +107,18 @@ contract Voucher is OwnableUpgradeable, IVoucher {
     ) external returns (bytes32 dealId) {
         // TODO add onlyAuthorized
         // TODO check expiration
-        uint256 appPrice = appOrder.appprice;
-        uint256 datasetPrice = datasetOrder.datasetprice;
-        uint256 workerpoolPrice = workerpoolOrder.workerpoolprice;
         VoucherStorage storage $ = _getVoucherStorage();
         IVoucherHub voucherHub = IVoucherHub($._voucherHub);
-        uint256 sponsoredAmount = voucherHub.debitVoucher(
-            $._type,
-            appOrder.app,
-            appPrice,
-            datasetOrder.dataset,
-            datasetPrice,
-            workerpoolOrder.workerpool,
-            workerpoolPrice
-        );
-        uint256 dealPrice = appPrice + datasetPrice + workerpoolPrice;
         address iexecPoco = voucherHub.getIexecPoco();
-        if (sponsoredAmount != dealPrice) {
-            // Transfer non-sponsored amount from the iExec account of the
-            // requester to the iExec account of the voucher
-            IERC20(iexecPoco).transferFrom(
-                requestOrder.requester,
-                address(this),
-                dealPrice - sponsoredAmount
-            );
-        }
+        uint256 sponsoredAmount = _debitVoucherAndTransferNonSponsoredAmount(
+            $._type,
+            voucherHub,
+            iexecPoco,
+            appOrder,
+            datasetOrder,
+            workerpoolOrder,
+            requestOrder
+        );
         dealId = IexecPoco1(iexecPoco).sponsorMatchOrders(
             appOrder,
             datasetOrder,
@@ -124,6 +127,45 @@ contract Voucher is OwnableUpgradeable, IVoucher {
         );
         $._sponsoredAmounts[dealId] = sponsoredAmount;
         emit OrdersMatchedWithVoucher(dealId);
+        return dealId;
+    }
+
+    /**
+     * Match orders boost on Poco. Eligible assets prices will be debited from the
+     * voucher if possible, then non-sponsored amount will be debited from the
+     * iExec account of the requester.
+     *
+     * @param appOrder The app order.
+     * @param datasetOrder The dataset order.
+     * @param workerpoolOrder The workerpool order.
+     * @param requestOrder The request order.
+     */
+    function matchOrdersBoost(
+        IexecLibOrders_v5.AppOrder calldata appOrder,
+        IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
+        IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
+    ) external onlyAuthorized onlyNotExpired returns (bytes32 dealId) {
+        VoucherStorage storage $ = _getVoucherStorage();
+        IVoucherHub voucherHub = IVoucherHub($._voucherHub);
+        address iexecPoco = voucherHub.getIexecPoco();
+        uint256 sponsoredAmount = _debitVoucherAndTransferNonSponsoredAmount(
+            $._type,
+            voucherHub,
+            iexecPoco,
+            appOrder,
+            datasetOrder,
+            workerpoolOrder,
+            requestOrder
+        );
+        dealId = IexecPocoBoost(iexecPoco).sponsorMatchOrdersBoost(
+            appOrder,
+            datasetOrder,
+            workerpoolOrder,
+            requestOrder
+        );
+        $._sponsoredAmounts[dealId] = sponsoredAmount;
+        emit OrdersBoostMatchedWithVoucher(dealId);
         return dealId;
     }
 
@@ -194,6 +236,55 @@ contract Voucher is OwnableUpgradeable, IVoucher {
     function _getVoucherStorage() private pure returns (VoucherStorage storage $) {
         assembly {
             $.slot := VOUCHER_STORAGE_LOCATION
+        }
+    }
+
+    /**
+     * @dev Debit voucher and transfer non-sponsored amount from requester's account.
+     *
+     * @param voucherTypeId The type Id of the voucher.
+     * @param voucherHub The voucher hub instance.
+     * @param iexecPoco The address of iExec Poco.
+     * @param appOrder The app order.
+     * @param datasetOrder The dataset order.
+     * @param workerpoolOrder The workerpool order.
+     * @param requestOrder The request order.
+     *
+     * @return sponsoredAmount The amount sponsored by the voucher.
+     */
+    function _debitVoucherAndTransferNonSponsoredAmount(
+        uint256 voucherTypeId,
+        IVoucherHub voucherHub,
+        address iexecPoco,
+        IexecLibOrders_v5.AppOrder calldata appOrder,
+        IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
+        IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
+    ) private returns (uint256 sponsoredAmount) {
+        uint256 appPrice = appOrder.appprice;
+        uint256 datasetPrice = datasetOrder.datasetprice;
+        uint256 workerpoolPrice = workerpoolOrder.workerpoolprice;
+
+        sponsoredAmount = voucherHub.debitVoucher(
+            voucherTypeId,
+            appOrder.app,
+            appPrice,
+            datasetOrder.dataset,
+            datasetPrice,
+            workerpoolOrder.workerpool,
+            workerpoolPrice
+        );
+        // TODO: Compute volume and set dealPrice = taskPrice * volume instead of curent dealPrice
+        uint256 dealPrice = appPrice + datasetPrice + workerpoolPrice;
+
+        if (sponsoredAmount != dealPrice) {
+            // Transfer non-sponsored amount from the iExec account of the
+            // requester to the iExec account of the voucher
+            IERC20(iexecPoco).transferFrom(
+                requestOrder.requester,
+                address(this),
+                dealPrice - sponsoredAmount
+            );
         }
     }
 }