+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNEXT
 - Use real poco address if available at deployment. (#17)
+- Match orders boost through voucher. (#16)
 - Use hardhat deploy. (#15)
 - Upload coverage reports to Codecov. (#14)
 - Clean some TODOs. (#13)